Add Angular example implementing a custom element to workspace
This is the same Angular code we use in angular-mvc-portlet but in its
regular format (ie: not tuned to be deployed from an MVC portlet).

Note that we disable the `test` script in `package.json` because it would
block gradle's build since it launches a browser to execute E2E points.